renamed sql_app package to web_apps
